refacotize aligns extensions into only one and cleanning
